requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

falcon
jsonschema
oslo.config
oslo.context
oslo.db
oslo.log
oslo.middleware
keystonemiddleware
PasteDeploy
pbr>=1.6
